+++ src/hooks/game_manager.cpp
@@ -2,7 +2,9 @@
 
 #include "gjgamelevel.hpp"
 
+#include <managers/settings.hpp>
 #include <util/debug.hpp>
+#include <util/ui.hpp>
 
 using namespace geode::prelude;
 
@@ -18,41 +20,22 @@ void HookedGameManager::returnToLastScene(GJGameLevel* level) {
 CCTexture2D* HookedGameManager::loadIcon(int iconId, int iconType, int iconRequestId) {
     if (m_fields->iconCache.contains(iconType)) {
         if (m_fields->iconCache.at(iconType).contains(iconId)) {
-            return m_fields->iconCache.at(iconType).at(iconId);
+            auto* tex = *m_fields->iconCache.at(iconType).at(iconId);
+            return tex;
         }
     }
 
-    auto* gm = GameManager::get();
-    auto* textureCache = CCTextureCache::sharedTextureCache();
-    auto* sfCache = CCSpriteFrameCache::sharedSpriteFrameCache();
-    CCTexture2D* texture = nullptr;
-
-    std::string sheetName = gm->sheetNameForIcon(iconId, iconType);
-    if (!sheetName.empty()) {
-        int key = gm->keyForIcon(iconId, iconType);
-
-        if (gm->m_loadIcon[key] < 1) {
-            texture = textureCache->addImage((sheetName + ".png").c_str(), false);
-            sfCache->addSpriteFramesWithFile((sheetName + ".plist").c_str());
-        } else {
-            texture = textureCache->textureForKey((sheetName + ".png").c_str());
-        }
-
-        auto& thatMap = gm->m_loadIcon2[iconRequestId];
-        auto existingIcon = thatMap[iconType];
-        if (existingIcon != iconId) {
-            gm->m_loadIcon[key]++;
-            if (existingIcon > 0) {
-                gm->unloadIcon(existingIcon, iconType, iconRequestId);
-            }
-
-            gm->m_loadIcon2[iconRequestId][iconType] = iconId;
-        }
-    }
+    CCTexture2D* texture = GameManager::loadIcon(iconId, iconType, iconRequestId);
 
     if (texture) {
         m_fields->iconCache[iconType][iconId] = texture;
     }
 
     return texture;
+}
+
+void HookedGameManager::reloadAllStep2() {
+    log::debug("cleared icon cache");
+    m_fields->iconCache.clear();
+    GameManager::reloadAllStep2();
 }