+++ src/hooks/game_manager.hpp
@@ -4,8 +4,9 @@
 #include <Geode/modify/GameManager.hpp>
 
 class $modify(HookedGameManager, GameManager) {
-    std::unordered_map<int, std::unordered_map<int, cocos2d::CCTexture2D*>> iconCache;
+    std::unordered_map<int, std::unordered_map<int, Ref<cocos2d::CCTexture2D>>> iconCache;
 
     void returnToLastScene(GJGameLevel* level);
     cocos2d::CCTexture2D* loadIcon(int iconId, int iconType, int iconRequestId);
+    void reloadAllStep2();
 };